LRNG - sysctls and /proc interface
The LRNG sysctl interface provides the same controls as the existing
/dev/random implementation. These sysctls behave identically and are
implemented identically. The goal is to allow a possible merge of the
existing /dev/random implementation with this implementation which
implies that this patch tries have a very close similarity. Yet, all
sysctls are documented at [1].

The sysctl implementation is only enabled if the existing random.c file
is not compiled.

Signed-off-by: Stephan Mueller <smueller@chronox.de>